WIP: Added framework for Listener (pushing to switch IDE to macOS)
