+++ src/Sentry.OpenTelemetry/SentryPropagator.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.Primitives;
 using OpenTelemetry;
 using OpenTelemetry.Context.Propagation;
 using Sentry.Extensibility;
+using Sentry.Internal.Tracing;
 
 namespace Sentry.OpenTelemetry;
